% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyBoundaries.R
\name{applyBoundaries}
\alias{applyBoundaries}
\title{applyBoundaries}
\usage{
applyBoundaries(box, data, complement = FALSE)
}
\arguments{
\item{box}{The result one box drawing algorithm (containing the boundaries).}

\item{data}{The underlying data.}

\item{complement}{Flag if the complement subset is needed.}
}
\value{
Subset of the data if there is only one run or else list of subsets.
}
\description{
Generates a subset of the data by the given box boundaries.
}
