% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactbox.R
\name{exactboxes}
\alias{exactboxes}
\title{Exactbox-Algorithm.}
\usage{
exactboxes(positiveTraining, negativeTraining, positiveTesting, negativeTesting,
  cSize, maxK, cExpand, timePerProblem = 4500, varType = "C", env = NULL)
}
\arguments{
\item{positiveTraining}{Positive training data.}

\item{negativeTraining}{Negative training data.}

\item{positiveTesting}{Positive test data.}

\item{negativeTesting}{Negative test data.}

\item{cSize}{Number of different weight for negative data point.}

\item{maxK}{Cluster size.}

\item{cExpand}{Parameter to control the number of boxes.}

\item{timePerProblem}{Time that you are willing to invest for a single MIP problem.}

\item{varType}{}
}
\value{
A list containing training TP, training FP, testing FTP and testing FP.
}
\description{
This code requires gurobi and the gurobi-library to solve the MIP.
}
\author{
Algorithm and Matlab-code by Cynthia Rudin and Siong Thye Go (see References). Implemented in R by Hendrik Pfaff
}
\keyword{Integer}
\keyword{Mixed}
\keyword{Programming,}
\keyword{box}
\keyword{classification}
