% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbox.R
\name{fastboxes}
\alias{fastboxes}
\title{Fastbox-Algorithm}
\usage{
fastboxes(positiveTraining, negativeTraining, positiveTesting, negativeTesting,
  cSize, idealK, beta)
}
\arguments{
\item{positiveTraining}{Positive training data.}

\item{negativeTraining}{Negative training data.}

\item{positiveTesting}{Positive test data.}

\item{negativeTesting}{Negative test data.}

\item{cSize}{The size that of weight for negative data to try. It will enumerate weight from 1/cSize, 2/cSize, etc... up to 1.}

\item{idealK}{The number of clusters.}

\item{beta}{The expansion parameter.}
}
\value{
A list containing training TP, training FP, testing TP, testing FP and the upper/lower ideal boundray.
}
\description{
Use the Fastbox-Algorithm for classification.
}
\author{
Algorithm and Matlab-code by Cynthia Rudin and Siong Thye Go (see References). Translation to R by Hendrik Pfaff
}
\keyword{box}
\keyword{classification}
