% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myAll.R
\name{myAll}
\alias{myAll}
\title{myAll}
\usage{
myAll(x)
}
\arguments{
\item{x}{Matrix of Data.}
}
\value{
Vector containing TRUE or FALSE.
}
\description{
R equivalent of the matlab function all(A, 2). Checks of a row of Data is nonzero or true.
}
