% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printBoundaries.R
\name{printBoundaries}
\alias{printBoundaries}
\title{printBoundaries}
\usage{
printBoundaries(box, data = NULL, nonRelevant = FALSE)
}
\arguments{
\item{box}{The result one box drawing algorithm (containing the boundaries).}

\item{nonRelevant}{Flag if non relevant boundaries should be printed.}

\item{colnames}{Names for each column/dimension of the data.}
}
\value{
void if everything works fine.  NULL if no boundaries are found.
}
\description{
Ouputs the boundaries for all found boxes.
}
