% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceObsByHemming.R
\name{reduceObsByHemming}
\alias{reduceObsByHemming}
\title{Reducing the negative Observations that are too different from the positive.}
\usage{
reduceObsByHemming(pos, neg, maxDist)
}
\description{
Reducing the negative Observations that are too different from the positive.
}
