% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Repmat}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{Matrix to be repeated.}

\item{m}{Number of repetitions  columnwise.}

\item{n}{Number of repititions rowwise.}
}
\value{
New Matrix with repeated values.
}
\description{
R equivalent of the matlab function repmat. Repeats a matrix column- and rowwise.
}
